spec: add initial build dir tests
The run_mock_project was moved into utils to help with this and given
some new functionality. It can now specify the command to run and the
arguments to run it with. As well as some basic assertions about the
types of its arguments.
